chore(public-calendar): clarify service_key docs and harden repo
- Inline comment: lowercase before pattern for mixed-case slugs
- Log filter.service_type only when None (not empty set)
- OpenAPI: distinguish calendar query param from discount body field
- list_public_offerings: docstring + defensive service_key.lower()
- Merge parser tests into test_public_events_api.py
